DX-914: Rename logger.rb to log_wrapper.rb
Rename `logger.rb` to `log_wrapper.rb` to avoid collisions with
`/usr/local/Cellar/ruby@2.7/2.7.4/lib/ruby/2.7.0/logger.rb`.
