+++ lib/kramdown-plantuml/diagram.rb
@@ -3,7 +3,7 @@
 require_relative 'version'
 require_relative 'theme'
 require_relative 'plantuml_error'
-require_relative 'logger'
+require_relative 'log_wrapper'
 require_relative 'executor'
 
 module Kramdown
