+++ lib/kramdown-plantuml/log_wrapper.rb
